+++ src/hooks/useFetch/useFetch.js
@@ -1,9 +1,57 @@
 import { useState, useEffect } from 'react';
-
-const useFetch = (url) => {
-	useEffect(() => {}, []);
-
-	return { response };
+import axios from 'axios';
+/**
+ * @param {string} url
+ */
+const useNextApiFetch = (url) => {
+	const [data, setData] = useState(null);
+	const [loading, setLoading] = useState(true);
+	const [error, setError] = useState(null);
+	const fetchData = async (url) => {
+		axios
+			.get(url)
+			.then((res) => {
+				setData(res.data);
+				setLoading(false);
+			})
+			.catch((error) => {
+				console.error(error);
+				setError(error);
+				setLoading(false);
+			});
+	};
+	useEffect(() => {
+		const currentUrl = window.location.origin;
+		fetchData(currentUrl + url);
+	}, [url]);
+	console.log({ data, loading, error });
+	return { data, loading, error };
 };
+/**
+ * @param {string} url
+ */
+const useFetch = (url) => {
+	const [data, setData] = useState(null);
+	const [loading, setLoading] = useState(true);
+	const [error, setError] = useState(null);
+	const fetchData = async (url) => {
+		axios
+			.get(url)
+			.then((res) => {
+				setData(res.data);
+				setLoading(false);
+			})
+			.catch((error) => {
+				console.error(error);
+				setError(error);
+				setLoading(false);
+			});
+	};
+	useEffect(() => {
+		fetchData(url);
+	}, [url]);
 
+	return { data, loading, error };
+};
+export { useNextApiFetch };
 export default useFetch;
